requirements.txt
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.3.0
configobj==5.0.9
configparser==7.2.0
dataclasses-json==0.6.7
ddgs==9.6.0
distro==1.9.0
et_xmlfile==2.0.0
etelemetry==0.3.1
faiss-cpu==1.12.0
fastapi==0.117.1
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.30
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langsmith==0.4.31
looseversion==1.3.0
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.3
openai==1.109.1
openpyxl==3.1.5
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pathlib==1.0.1
pillow==11.3.0
pipdeptree==2.28.0
primp==0.15.0
propcache==0.3.2
prov==2.1.1
puremagic==1.30
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydot==4.0.1
PyMuPDF==1.26.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.3
rdflib==7.2.1
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
SQLAlchemy==2.0.43
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
traits==7.0.2
transformers==4.56.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
yarl==1.20.1
zstandard==0.25.0
